chore: remove BPD, SiciliaVola and Bonus Vacanze features
