+++ package.json
@@ -6,7 +6,6 @@
   "main": "app.js",
   "api_backend_specs": "https://raw.githubusercontent.com/pagopa/io-backend/v13.25.1-RELEASE/api_backend.yaml",
   "api_public_specs": "https://raw.githubusercontent.com/pagopa/io-backend/v13.25.1-RELEASE/api_public.yaml",
-  "api_bonus_vacanze": "https://raw.githubusercontent.com/pagopa/io-backend/v13.25.1-RELEASE/api_bonus.yaml",
   "api_cgn": "https://raw.githubusercontent.com/pagopa/io-backend/v13.25.1-RELEASE/api_cgn.yaml",
   "api_cgn_merchants": "https://raw.githubusercontent.com/pagopa/io-backend/v13.25.1-RELEASE/api_cgn_operator_search.yaml",
   "api_cgn_geo": "https://raw.githubusercontent.com/pagopa/io-backend/here_geoapi_integration/api_geo.yaml",
@@ -15,14 +14,7 @@
   "api_pagopa": "https://raw.githubusercontent.com/pagopa/io-app/master/assets/paymentManager/spec.json",
   "pagopa_cobadge_configuration": "https://raw.githubusercontent.com/pagopa/io-services-metadata/1.0.29/pagopa/cobadge/abi_definitions.yml",
   "pagopa_privative_configuration": "https://raw.githubusercontent.com/pagopa/io-services-metadata/1.0.29/pagopa/privative/definitions.yml",
-  "api_bpd_citizen": "https://raw.githubusercontent.com/pagopa/io-services-metadata/1.0.29/bonus/specs/bpd/citizen.json",
-  "api_bpd_citizen_v2": "https://raw.githubusercontent.com/pagopa/io-services-metadata/1.0.29/bonus/specs/bpd/citizen_v2.json",
-  "api_bpd_payment": "https://raw.githubusercontent.com/pagopa/io-services-metadata/1.0.29/bonus/specs/bpd/payment.json",
-  "api_bpd_award": "https://raw.githubusercontent.com/pagopa/io-services-metadata/1.0.29/bonus/specs/bpd/award.json",
-  "api_bpd_winning_transactions": "https://raw.githubusercontent.com/pagopa/io-services-metadata/1.0.29/bonus/specs/bpd/winning_transactions.json",
-  "api_bpd_winning_transactions_v2": "https://raw.githubusercontent.com/pagopa/io-services-metadata/1.0.29/bonus/specs/bpd/winning_transactions_v2.json",
   "api_eu_covid_cert": "https://raw.githubusercontent.com/pagopa/io-backend/v13.25.1-RELEASE/api_eucovidcert.yaml",
-  "api_sicilia_vola": "assets/SiciliaVola.yml",
   "api_cdc": "https://raw.githubusercontent.com/pagopa/io-app/master/assets/CdcSwagger.yml",
   "api_fci": "https://raw.githubusercontent.com/pagopa/io-backend/v13.25.1-RELEASE/api_io_sign.yaml",
   "api_pn": "https://raw.githubusercontent.com/pagopa/io-backend/v13.25.1-RELEASE/api_pn.yaml",
@@ -53,18 +45,10 @@
     "generate:pagopa-walletv3-definitions": "rimraf generated/definitions/pagopa/walletv3 && mkdir -p generated/definitions/pagopa/walletv3 && gen-api-models --api-spec $npm_package_api_pagopa_walletv3 --out-dir ./generated/definitions/pagopa/walletv3",
     "generate:pagopa-ecommerce-definitions": "rimraf generated/definitions/pagopa/ecommerce && mkdir -p generated/definitions/pagopa/ecommerce && gen-api-models --api-spec $npm_package_api_pagopa_ecommerce --out-dir ./generated/definitions/pagopa/ecommerce",
     "generate:pagopa-api": "rimraf generated/definitions/pagopa && mkdir -p generated/definitions/pagopa && gen-api-models --api-spec $npm_package_api_pagopa --out-dir ./generated/definitions/pagopa",
-    "generate:bonus-vacanze-definitions": "rimraf generated/definitions/bonus_vacanze && mkdir -p generated/definitions/bonus_vacanze && gen-api-models --api-spec $npm_package_api_bonus_vacanze --out-dir ./generated/definitions/bonus_vacanze ",
     "generate:cgn-definitions": "rimraf generated/definitions/cgn && mkdir -p generated/definitions/cgn && gen-api-models --api-spec $npm_package_api_cgn --out-dir ./generated/definitions/cgn ",
     "generate:cgn-merchants-definitions": "rimraf generated/definitions/cgn/merchants && mkdir -p generated/definitions/cgn/merchants && gen-api-models --api-spec $npm_package_api_cgn_merchants --out-dir ./generated/definitions/cgn/merchants ",
     "generate:cgn-geo-definitions": "rimraf generated/definitions/cgn/geo && mkdir -p generated/definitions/cgn/geo && gen-api-models --api-spec $npm_package_api_cgn_geo --out-dir ./generated/definitions/cgn/geo ",
-    "generate:bpd-citizen": "rimraf generated/definitions/bpd/citizen && mkdir -p generated/definitions/bpd/citizen  && gen-api-models --api-spec $npm_package_api_bpd_citizen --out-dir ./generated/definitions/bpd/citizen --no-strict",
-    "generate:bpd-citizen-v2": "rimraf generated/definitions/bpd/citizen-v2 && mkdir -p generated/definitions/bpd/citizen-v2  && gen-api-models --api-spec $npm_package_api_bpd_citizen_v2 --out-dir ./generated/definitions/bpd/citizen-v2 --no-strict",
-    "generate:bpd-payment": "rimraf generated/definitions/bpd/payment && mkdir -p generated/definitions/bpd/payment  && gen-api-models --api-spec $npm_package_api_bpd_payment --out-dir ./generated/definitions/bpd/payment --no-strict",
-    "generate:bpd-award": "rimraf generated/definitions/bpd/award && mkdir -p generated/definitions/bpd/award  && gen-api-models --api-spec $npm_package_api_bpd_award --out-dir ./generated/definitions/bpd/award --no-strict",
-    "generate:bpd-winning-transactions": "rimraf generated/definitions/bpd/winning_transactions && mkdir -p generated/definitions/bpd/winning_transactions  && gen-api-models --api-spec $npm_package_api_bpd_winning_transactions --out-dir ./generated/definitions/bpd/winning_transactions --no-strict",
-    "generate:bpd-winning-transactions-v2": "rimraf generated/definitions/bpd/winning_transactions/v2 && mkdir -p generated/definitions/bpd/winning_transactions/v2  && gen-api-models --api-spec $npm_package_api_bpd_winning_transactions_v2 --out-dir ./generated/definitions/bpd/winning_transactions/v2 --no-strict",
     "generate:eu-covid-cert-definitions": "rimraf generated/definitions/eu_covid_cert && mkdir -p generated/definitions/eu_covid_cert && gen-api-models --api-spec $npm_package_api_eu_covid_cert --out-dir ./generated/definitions/eu_covid_cert --no-strict",
-    "generate:sicilia-vola-definitions": "rimraf generated/definitions/siciliaVola && mkdir -p generated/definitions/siciliaVola && gen-api-models --api-spec $npm_package_api_sicilia_vola --out-dir ./generated/definitions/siciliaVola --no-strict",
     "generate:cdc-definitions": "rimraf generated/definitions/cdc && mkdir -p generated/definitions/cdc  && gen-api-models --api-spec $npm_package_api_cdc --out-dir ./generated/definitions/cdc --no-strict",
     "generate:fci-definitions": "rimraf generated/definitions/fci && mkdir -p generated/definitions/fci  && gen-api-models --api-spec $npm_package_api_fci --out-dir ./generated/definitions/fci --no-strict",
     "generate:pn-definitions": "rimraf generated/definitions/pn && mkdir -p generated/definitions/pn && gen-api-models --api-spec $npm_package_api_pn --out-dir ./generated/definitions/pn --no-strict --request-types --response-decoders",
